fix and move utils
